Re-think the logging API, clearly separating the logger transport with the logger facade. Introduce a Logger interface, extending the Console interface, that will be used within the sdk by the shared core.
